+++ src/main/java/de/mossgrabers/convertwithmoss/format/music1010/bento/BentoDetector.java
@@ -30,6 +30,7 @@ import de.mossgrabers.convertwithmoss.core.detector.DefaultPerformanceSource;
 import de.mossgrabers.convertwithmoss.core.model.IEnvelope;
 import de.mossgrabers.convertwithmoss.core.model.IFilter;
 import de.mossgrabers.convertwithmoss.core.model.IGroup;
+import de.mossgrabers.convertwithmoss.core.model.IMetadata;
 import de.mossgrabers.convertwithmoss.core.model.ISampleData;
 import de.mossgrabers.convertwithmoss.core.model.ISampleLoop;
 import de.mossgrabers.convertwithmoss.core.model.ISampleZone;
@@ -210,7 +211,7 @@ public class BentoDetector extends AbstractDetector<MetadataSettingsUI>
     }
 
 
-    private Optional<IInstrumentSource> parseMultisample (final File multiSampleFile, final String defaultName, final Element trackElement, final Element instElement, final List<Element> assetElements, final String basePath)
+    private Optional<IInstrumentSource> parseMultisample (final File sourceFile, final String defaultName, final Element trackElement, final Element instElement, final List<Element> assetElements, final String basePath)
     {
         final Element trackParamsElement = XMLUtils.getChildElementByName (trackElement, Music1010Tag.PARAMS);
         final Element instParamsElement = XMLUtils.getChildElementByName (instElement, Music1010Tag.PARAMS);
@@ -223,9 +224,9 @@ public class BentoDetector extends AbstractDetector<MetadataSettingsUI>
 
         final File pathPrefixFile = new File (pathPrefix);
         final String name = pathPrefixFile.getName ();
-        final String [] parts = AudioFileUtils.createPathParts (multiSampleFile.getParentFile (), this.sourceFolder, name);
+        final String [] parts = AudioFileUtils.createPathParts (sourceFile.getParentFile (), this.sourceFolder, name);
 
-        final DefaultMultisampleSource multisampleSource = new DefaultMultisampleSource (multiSampleFile, parts, name, AudioFileUtils.subtractPaths (this.sourceFolder, multiSampleFile));
+        final DefaultMultisampleSource multisampleSource = new DefaultMultisampleSource (sourceFile, parts, name, AudioFileUtils.subtractPaths (this.sourceFolder, sourceFile));
         final IGroup group = new DefaultGroup ("Group");
         multisampleSource.setGroups (Collections.singletonList (group));
 
@@ -264,7 +265,9 @@ public class BentoDetector extends AbstractDetector<MetadataSettingsUI>
         parseEffects (instParamsElement, multisampleSource);
         readVelocityModulators (instElement, group);
 
-        this.createMetadata (multisampleSource.getMetadata (), this.getFirstSample (multisampleSource.getGroups ()), parts);
+        final IMetadata metadata = multisampleSource.getMetadata ();
+        this.createMetadata (metadata, this.getFirstSample (multisampleSource.getGroups ()), parts);
+        this.updateCreationDateTime (metadata, sourceFile);
         return Optional.of (new DefaultInstrumentSource (multisampleSource, midiChannel));
     }